made two more styled files to make the Cover.styled.jsx smaller and readable
